Enable project-level control of web-based commit signing
This change introduces support for enabling web-based commit signing
at the project level, as part of making it opt-in for SaaS.
The web_based_commit_signing_enabled parameter can be updated
only if the group-level namespace setting is not enabled.

When the group-level setting is enabled,
all projects under the group inherit it and cannot override it.

For personal projects (no group),
the project-level setting can be freely updated.

The default value is false, and the project value is cascaded
from the group setting if one exists.

Related issue: https://gitlab.com/gitlab-org/gitlab/-/issues/540914

Changelog: changed
EE: true