Add after_validation to ensure oid is updated
We need to duplicate the oid information from lfs_objects to
lfs_objects_project records. We've previously added the oid column, and
this MR will add an after_validation to ensure that the value is copied
over for new and updated records; this will ease our next step,
backfilling the data.