Remove misleading `//` which would actually break any expression
Based on testing, it's clear that the value in the field is used
directly in whatever Regex test is being performed on the title.

If a user includes the leading and trailing `/` in the value as is
implied by the placeholder, no expression works (unless you also
include `/` at the beginning and end of your MR title).

This is a more complicated expression, but since we can't include
the two `/`, it demonstrates a working regular expression whereas
simply `title` (without the `/`) could imply that only text is matched.