Reduce global database query count
Updates the Duo Chat template and super sidebar helper to
reduce the amount of database queries that happen globally.
For pages that aren't in the group or project context we won't render
Duo Chat so we don't need to render the chat template on the page.
The attribute in the super sidebar is no longer used so that can be removed
to reduce the query count more.

Locally on the merge request dashboard (with the `merge_request_dashboard` flag
enabled) this reduced the query count by 22.
The queries are relatively quick queries so won't slow the page down,
but because we don't have Duo Chat on these root pages we can safely
remove these queries and reduce the overall database queries.