Pass PATs scopes from backend to frontend in new UI
In the new UI for personal access tokens (vue_shared access tokens), we
replace hardcoded scopes in the frontend for values generated in the
backend (with pre-existent logic). We pass the values from the backend
to the frontend via a data attribute.

The only addition is that the scopes are now sorted from low to height
level of permission.

Changelog: changed