Document GitLab 18.0 issues with mixing FIPS and non-FIPS Docker image
The GitLab 18.0 Docker image upgraded the base image from Ubuntu 22.04
to 24.04. When the image is run on a FIPS host, users will encounter
`PRNG is not seeded` when GitLab attempts to run the OpenSSH server.

In https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/8450
we added support for setting the `OPENSSL_FORCE_FIPS_MODE=0`
environment variable as a workaround.

This commit documents the limitation of mixing FIPS and non-FIPS
system and how to work around the issue with this environment variable.