Remove outdated test
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/516801

**Problem**

The test was introduced in
https://gitlab.com/gitlab-org/gitlab/-/commit/bb4e47a812d159ac8a4b63d3e2485dfff29260b0
to remove explicitly metadata from UpdateAllMirrorsWorker context.

However, later this logic was extracted from UpdateAllMirrorsWorker:
https://gitlab.com/gitlab-org/gitlab/-/commit/c310003df95e65c64f84081d62cdc223e9a3227e.

We already have a test that verifies that project, namespace and user
are removed from the context:
https://gitlab.com/gitlab-org/gitlab/-/blob/c310003df95e65c64f84081d62cdc223e9a3227e/spec/workers/concerns/cronjob_queue_spec.rb#L25.

**Solution**

The test can be removed because this functionality is already covered by
other test.