Remove DesiredConfigGenerator versioned call
From response_payload_builder.rb

Versioning is removed from DesiredConfigGenerator in favor of freezing
kubernetes config during the create step of workspace. As a result this
code is not needed anymore. Older workspaces will use the config that is
already generated for them when they were created thereby eliminating
the need for the DesiredConfigGenerator to be versioned.

The version column and other model related changes will be done in a
separate MR later.