Improve project deletion error message
We prevent deleting or renaming a project when the project being deleted
still has container registry images, because otherwise the container
registry data will not get cleaned up. This can lead to security issues
when the old project name is used by a new project.

The current error message does not guide the user on how to address the
error so that the project deletion can proceed. This MR changes the
error message to provide more guidance.

Changlog: fixed