Add new rate limit for authenticated Git HTTP
This commit adds three new fields to the rate_limits jsonb column that
are related to rate limiting authenticated Git HTTP requests.

It adds the new rate limits for authenticated Git HTTP requests to the
throttle definitions and makes use of them in Rackt Attack. It checks if
the setting is enabled and if the path is a git path. It leaves out the
check against `unauthenticated` since this rate limit is for
authenticated requests.

This commit makes the authenticated Git HTTP rate limits setting visible
in the network settings page available to admins.

UI changes and backend checks have been placed behind a new feature flag
named git_authenticated_http_limit that is disabled by default.