Update `vulnerability_export` API to accept `report_data`
Context
------------------

In a [previous MR][0], we added a `jsonb` column to the
`vulnerability_exports`. This field is used to store assets from the
frontend, to be used in report rendering.

This change introduces the first usasge of this `jsonb` column. In
this instance, we are going to store an SVG that has been computed in
the frontend. Future usages may store json data blobs or other non-svg
data.

This change
------------------

Here, we are taking an SVG of the project-level
["vulnerabilities over time"][1] table, rendered in the web UI, as an
asset to be rendered in the PDF report.

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/192895
[1]:https://docs.gitlab.com/user/application_security/security_dashboard/#project-security-dashboard

---

epic: https://gitlab.com/groups/gitlab-org/-/epics/16989
Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/524058
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/193143

Changelog: changed
EE: true