Improves performance of tab switching on merge request dashboard
The tabs on the merge request use lazy rendered GitLab UI tabs
which stop data being from being fetched from a tab that isn't being viewed.
However if a user visit the `Merged` tab and then switches back to `Active`
there is a small render blocking delay because of Vue re-creating all
the components again.
With this change we cache what tabs have been visited and disable the tabs
lazy rendering for these tabs.

With this change switching between tabs is now near instant and will only be
slowed down by the browsers ability to quickly toggle the display
property and render the elements on screen.

A user with 140 merge requests on the page (this could be larger if they have
paginated through a list) it will remove a render blocking delay of around 400ms, this delay will be larger on less powerful machines.