Change `protect_merge_request_pipelines` column default value to true
Initially, we added a new column `protect_merge_request_pipelines` to
the `project_settings` table with a default value of `false` although we
wanted new projects created after this migration to have the default
value of `true`. This was done to ensure that existing
CI pipelines would not be affected by this change. Refer to this
migration https://gitlab.com/gitlab-org/gitlab/-/merge_requests/187342
for more details.

To have the correct default value for newly created projects,
we set the default value of the `protect_merge_request_pipelines`
column to `true` in the application code here -
https://gitlab.com/gitlab-org/gitlab/-/blob/master/app/services/projects/create_service.rb#L34
by setting params[:protect_merge_request_pipelines] = true

This commit reflects the default value set in the application code
by changing the default value of the `protect_merge_request_pipelines`
column to `true` at the database level.

https://gitlab.com/gitlab-org/gitlab/-/issues/547482

https://docs.gitlab.com/development/database/avoiding_downtime_in_migrations/#changing-column-defaults

Changelog: changed