When entering zen mode, set fixed MR top bar to static
The fixed position top bar on an MR is in an earlier stacking context
than the zen mode fullscreen element, so it will appear on top of it.

This MR temporarily sets the top bar to have a static postion while zen
mode is on, so that it does not overlap the zen mode element.

Changelog: fixed