Add v2 approval rules data mapper
This commit adds a utility class that can take a v1 approval rule and
write it in the data base using the v2 models. It's not currently in use
in any of the coding paths at the time of this commit, but it can be
useful in development try out migrating data.

In the future we may we want to rewrite this using SQL instead of ruby
models so that we can start charting a path towards a backfill
migration.