Add the hugo shortcodes to prevent confusion
I followed this and failed to notice that we need to add the shortcodes. My editor gave me a warning but as a non-tech-writing user I wasn't familiar with what hugo shortcodes were. It took me a few minutes to realise that the examples below included the shortcodes and I asked the docs team for more information.

This change should make it more obvious to users which are unfamiliar with hugo.