Allow configuring image for Duo Workflow remote execution
You can read more about Duo Workflow at
https://handbook.gitlab.com/handbook/engineering/architecture/design-documents/duo_workflow/
. The "Remote Execution" feature allows you to trigger a workflow to run
inside a CI job. Previously this feature just used a hardcoded docker
image as the base image to run the CI job in. That is not particularly
useful as it might not have the right development tooling configured in
it.

Note that this features is only partially implemented and only available
via the API for now. It's not available to external users but it can
still be tested internally but it doesn't do much.

This MR introduces the ability to pass the `image` query parameter when
creating a workflow via the API. This `image` is persisted with the
workflow and is then used when constructing the CI job which runs the
workflow for Remote Execution.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/549763