Fix DateRangesDropdown spec under Vue 3
The spec and component assumed that the `v-model` directive would update
the local `selectedValue` property first, then call the `selected` event
handler.

That is true under Vue 2, but apparently not under Vue 3, where their
order is reversed, producing the incorrect result.

The solution is to not rely on that implicit behaviour of `v-model`, and
have the `selected` handler update the property explicitly.