Fix preserve markdown logic for lists
In RTE, if coming from the PTE with a bullet list
and changing it to another list type like task list or
numeric list, the change wouldn't apply. This was because
even though the list type changed, all children of the list
were unchanged. This commit changes the logic to check if
the parent's list type has changed, the children should
also be marked as changed.

Changelog: fixed