Specify which notification event to enable in docs
We have this section describing how to receive notifications when an
MR is created which you would be eligible to approve but it currently
says `select this event` without describing which event `this` refers
to.

This change specifies that the user should select the
`Merge request you're eligible to approve is created` event.