User OrganizationUserDetail in mentions, autocomplete, participants
Part 3 of our effort to allow internal bot users to be scoped
per-organization. This allows users with an
Organizations::OrganizationUserDetail record to be @ mentioned
within projects or groups in an organization by the alias rather than by
their username. Noteables mentioning these users will link to the
orignal username, not the alias, but will display with the alias.

Changelog: added