Allow DataTable fields to be customized
Adds support for the `fields` property in the datatable
`options` for configuring the fields to render. Falls back
to a default set of fields based on the first row of data
if the fields property is not specified.

Update the related storybook to include usage
examples.

Changelog: added
EE: true