Show vulnerability severiy only if an override is present
To support vulnerability severity overrides, we prioritize the
vulnerability severity over the security finding severity. However, this
logic isn't correct since severity can change, which will cause the
incorrect severity to be shown if the report contains a new severity and
the vulnerability does not have a severity override set. To fix this, we
change the logic so that vulnerability severity is only used if there is
an override.

Changelog: fixed
EE: true