Adjust placement of "Milestone can be closed" alert
When all the issues in a milestone are closed, a success alert shows on
the milestone page informing the user that the milestone can be closed.

This MR adjusts the placement of that alert to better comply with
Pajamas guidance. It also changes the logic so that this alert is only
shown to users who can actually take action to close a milestone.

Changelog: fixed