Add Max number of files setting for zoekt indexable project
Add an admin setting to set the max number of files in a project that
can be indexed for zoekt search. With this limit, we will not exhaust
resources to try projects which can not be indexed due to file count.

Changelog: other
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/194428
EE: true

# Conflicts:
#	ee/app/models/ee/application_setting.rb
#	ee/app/services/search/zoekt/task_serializer_service.rb
#	ee/spec/helpers/ee/application_settings_helper_spec.rb
#	ee/spec/models/search/zoekt/settings_spec.rb