Build#debug_mode? to return true when degenerated
This is a first iteration in preparation for potential
truncation of builds_metadata partition 100. The logic
needs to be updated when we migrate intrinsic metadata
to p_ci_builds. See
gitlab.com/gitlab-org/gitlab/-/merge_requests/194954#note_2574776849

Currently, when a job's build_metadata record is missing,
debug_mode? defaults to false, which
allows potentially sensitive debug trace details
to be viewable by unauthorized users.
We update the method so that it will be true for jobs that
get their metadata record deleted.

Changelog: changed