Fix code owner validation for roles
Fixes https://gitlab.com/gitlab-org/gitlab/-/issues/547265

In 18.0 we improved the validation logic for code owners. As part of
these changes we improved how we detect roles within the CODEOWNERS
files.

When we made this change it was overlooked that the regex to find the
roles allow singular and plural roles as well as case-insensitivity. As
this was overlooked. Prior to this change we would first convert the
strings to their integer values and then when validating the section
headers we would only check if the downcase singular form was provided
i.e. @@developer passed but @@DEVELOPER or @@developers would fail.

This change completely changes how we return the roles from the
reference extractor. Previously we would only return a single value for
each role found even if two variants have been specified i.e.

```codeowners
* @@DEVELOPER
CODEOWNERS @@DEVELOPERS
```

would only return `@@developer`.

After this change we now return the provided values `["@@DEVELOPER",
"@@DEVELOPERS"]`. This allows us to correctly highlight which owners are
not valid.

Changelog: fixed
EE: true