Add project vulnerabilities history to the pdf export
In a [previous MR][0], we started consuming SVG data from the front
end ["vulnerabilities over time"][1] dashboard with the idea of
rendering it in the pdf report

This change
-------------------

For the pdf component there are 3 main elements

1. The report title
2. The SVG from the front end
3. The key for the chart

The SVG
====================

There is some pre-processing of the svg required:[^1]

1. remove the svg from the `"data:image/svg+xml;charset=UTF-8, ... "` wrapper
2. un-urlencode the string
3. replace css variables with hard-coded colors[^2]

The chart key
====================

The front end is not able to send the key, as it is rendered in
html. So, we implement a basic key with prawn's native components.

>>> [!important]
The web UI uses the chart key to toggle the different
severity values in the chart. We are just showing a static list for
now. We haven't yet decided how we want to render "toggled off"
values. So, for now we have a static key
>>>

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/193143
[1]:https://docs.gitlab.com/user/application_security/security_dashboard/#project-security-dashboard

[^1]: We will investigate in followup work if there are any knobs on
      the charting library that we can use to receive the SVG in a better
      state.
[^2]: `prawn-svg` cannot render css variables

---

epic: https://gitlab.com/groups/gitlab-org/-/epics/16989
related to: https://gitlab.com/gitlab-org/gitlab/-/issues/524056
Changelog: added
EE: true