Provide fallback background-color for prefers-reduced-transparency
The `translucent-glass-background` mixin provides a background-color for
the `prefers-reduced-transparency: no-preference` media query, but does
not provide a fallback/default background-color. This means that users
who set `prefers-reduced-transparency` to `true` do not have a
background-color set which results in broken UI when fixed headers
display over other contents. This change leverages the CSS cascade to
provide a opaque fallback value using the `$bgColor` argument value

Changelog: fixed