wiki: Check if empty before querying files
Whenever the Repository home page is loaded, the ProjectPresenter will
check if the project's Wiki has a home page. This is used to determine
how the Wiki link should be rendered -- either as an "Add Wiki" link or
a button to navigate to the existing Wiki.

This call will eventually reach the find_matched_files method and then
result in a Gitaly call to the SearchFilesByName RPC.

If this happens on a project that has an empty Wiki repository (i.e. a
project which does not have any Wiki pages), SearchFilesByName will
fail. Internally, Gitaly will log a
"fatal: Not a valid object name <branch>" error, where <branch> is the
default branch of the Wiki repository. Rails however will simply
swallow this error and return no results, allowing the ProjectPresenter
to display the "Add Wiki" link.

The logs for `gprd` show that we're producing over one million of these
superfluous error logs per week. On top of wasting disk space, the
errors pollute real errors and make it difficult to monitor Gitaly in
production.

Add a check to repository.has_visible_content? before making any search
queries for individual files. has_visible_content? simply calls the
HasLocalBranches Gitaly RPC, which will return false if the repository
exists but is empty.