Security reports can only contain one scan or scanner
While reading `lib/gitlab/ci/reports/security/report.rb`, I was
extremely puzzled as to why it has multiple scans and scanners. Security
reports may only contain a single scan or scanner and this has been true
for many years. Most of this code was last edited 6 years ago, so I can
only conclude that this is a relic of another era when we had multiple
scanners in a single report. This refactor changes the `Report` class
and all callers to use simple `scan` and `scanner` accessors.