Add worker to delete expired and revoked OAuth Access Tokens
The `oauth_access_tokens` table size has grown substantially and does
not have a proper cleanup mechanism. There are over a 100 million
expired and revoked tokens each.
This commit contains changes to delete the existing invalid tokens.
Once the table size is under control, we need a subsequent change to
add a cron job that deletes expired and revoked tokens on a periodic
basis.

https://gitlab.com/gitlab-org/gitlab/-/issues/521855

Changelog: changed