Update authentications support in the maven virtual registry
Update the authentication support for the maven virtual registry endpoints
that are listed as part of the public API.
These endpoints need to support the authentication methods and transports
described in https://docs.gitlab.com/api/rest/authentication.

This does not impact the endpoints used by maven clients since they use a
specific set of authentications.