Make initials-only rails avatar component default to aria-hidden
The Rails Avatar component currently uses the item name as a default alt
text if no alt text is provided, but this makes no sense when there is
no image.

This MR adjusts the default behavior so that by default, initial-only
rails avatars will have no alt text and will be aria-hidden.

This MR also adjusts the usage of rails avatar components on Group or
Project homepages so that the avatar that immediately precedes the name
of the group or project will also have aria-hidden=true, since even if
it does have an image, reading the group or project name aloud here will
be useless to the screen reader user, who will hear the group or project
name read aloud again by the text that follows the avatar.

Changelog:fixed