Fix code coverage not showing up jobs running in `rspec fail-fast`
Previously if you had code that was exercised by tests that only ran
in the `rspec fail-fast` jobs, the code coverage data was not being
downloaded and merged into the final report. As a result, when GitLab
displayed merge request diffs with visual markers of code coverage,
the code would show up as untested code with red bars.

Add both the `rspec fail-fast` and `rspec-ee fail-fast` jobs to the
optional list of artifacts to collect. Note that these jobs only run
in EE, so we can add them to `rspec:artifact-collector ee remainder`.