Deduplicate label token results
In group contexts, multiple identical labels might be present in the
labels search dropdown (each in a separate project). Clicking on any of
them will filter on all of them (i.e., results that match any of the
various project labels that have the same name), so we might as well
deduplicate the list and only show one of them.

Changelog: changed