Skip unenforceable policy rules notification for child pipelines
Triggering the notification from child pipelines may cause noise
on the merge request if the scanners are only present in the
parent pipeline.

It's enough to trigger the notification when the parent
pipeline finishes, because it already checks all related
pipelines to look for the security artifacts.

Changelog: fixed
EE: true