Disable db:invalidate-old-pipeline-results job for forks
This job currently fails for merge requests from forks because
the source branch does not exist in the canonical repository.

This is a temporary fix to allow community contributions that
contain database changes to be merged, in a follow-up we will
update the job definition to handle these and re-enable it.