Check in CI that GDK can update after merging MR
This adds a `gdk-update` CI job that updates a GDK installation using
the `gdk update` command.

This ensures that after merging an MR, development environments should
be able to update to and run on the latest changes.

That is a way to shield engineers from instabilities that won't impact
production but, for example, break development databases or
dependencies.