Fix protected variables access in protected MR pipelines
In a previous merge request
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/188008
we introduced a change that allowed protected variables to be accessed
in merge request pipelines. However, this change was not complete.
In the previous implementation, we only set the pipeline as protected.
But we missed to update the logic that actually exposes protected
variables to the pipeline. This is because the code was NOT DRY and the
logic was duplicated in the `Gitlab::Ci::Variables::Builder` class.

This commit fixes the issue by resuing the existing logic that was
implemented in the original merge request.

Issue - https://gitlab.com/gitlab-org/gitlab/-/issues/552338

Changelog: changed