Show due date in second row to not take too much width
Most to-do items are not "due" anyway, and if they are, showing it in
its own line even helps readability.

But main reason is to give the todo content more room to breath.

For mobile (< sm) where the timestamps sit on their own row, we keep
the current behavior of showing them in one row with a middot.