Implement the first iteration of the activity feed widget
This implements the first iteration of the homepage's activity feed
widget. Provided that we don't yet have proper API endpoints to fetch
the necessary data, we currently piggy back on the internal endpoint
that the profile page uses. That endpoint returns raw HTML which we can
render as is within the widget, with minor style adjustments.