Namespacing dependency graph rebuild cache key
Currently, the cache key is just the SHA256 digest with no namespacing.
This could be problematic (nothing to stop a collision with another
project or feature usage, even if unlikely) but also makes it impossible
to introspect _what_ the cache key is linked to. This change aims to
make this clearer.

Changelog: changed
EE: true