Ensure blob content viewer works when navigate from blob to tree app
This is part of a bigger effort to implement a file tree browser (FTB)
in the repository.

Since we load the blob app when navigate from blob
page, we need to include `<router-view/>`
so vue routing works. This fix the issue by
mounting ~/repository/components/app.vue component,
which contains <router-view /> .