OpenAPI v3: Improve path parameter (project) id
**Summary**

The purpose of this commit is to simplify the OpenAPI document by following
the DRY pattern and using a single component `ProjectIdOrPath`
for every path parameter "id".

**Current Issues**

**Description Inconsistencies**

- The ID or URL-encoded path of the project
- The ID or URL-encoded path of the project owned by the authenticated user
- The ID of the project (where type is integer)

**Type Inconsistencies**

- Some parameters are described with type 'string'
- Some parameters are described with type 'integer'

**Proposed Changes**

To address these inconsistencies, I propose using the `anyOf` structure
to allow for both string and integer types. This approach aligns with
the markdown documentation (cf example below)
and provides clarity on the acceptable parameter types:

```yaml
schema:
  anyOf:
    - type: string
      example: gitlab-org/gitlab
    - type: integer
      example: 278964
```

**Examples**

For example, for the endpoint `GET /projects/{id}/badges/{badge_id}`,
the related markdown documentation clearly states the possibilities:

[Get a badge of a project](https://docs.gitlab.com/api/project_badges/#get-a-badge-of-a-project)

**Additional Information**

- I have also introduced a link to the
[namespaced-paths](https://docs.gitlab.com/api/rest/#namespaced-paths)
section in the parameter description to provide further clarity.