Refactor returns in begin blocks
Refactors a couple instances of returns in begin/end blocks
per rubocop rules. Adds an additional test for blank refs
previously handled by a return in a begin/end block.

Changes the behaviour slightly in merge_request.rb, where
previously calls to all_commit_shas would only be memoized
when persisted? == true. Searching the codebase for calls to
all_commit_shas shows that this is not an issue, as calls are
not repeatedly made or expect the value of all_commit_shas to
change if the value of persisted? changes.