Fix and un-quarantine group hook spec
This spec was marked as flaky and quarantined. Looking closer, it
appeared to be in fact reliably failing, as it expected an error
message in the incorrect format.

This MR updates the format to the error message produced used on
`ee/app/controllers/groups/hooks_controller.rb:25`:

```ruby
format(_('Hook execution failed. %{error}'), error: result.message)
```