Skip cyclic paths in Sbom::PathFinder
Skip over cyclic paths in the Finder algorithm.
This is done to prevent unrelated cyclic paths
from being added to the list of paths for an
occurrence. Since we are starting at root and
going down if we stop when we find a cycle
there is no way to know if we are on a path
leading to the target occurrence.

Changelog: other
EE: true