Run Create::DesiredConfig::Main in shadow with OldDesiredConfigGenerator
`DesiredConfigFetcher` is responsible for fetching desired_config. It
tries to fetch the desired_config from the database but in case it
cannot find it there, it calls
Create::DesiredConfig::Main.main function.

`ConfigToApplyBuilder` is responsible for injecting secrets and filter
out redundant objects from desired_config. It uses
DesiredConfigFetcher class to fetch the desired_config.

`ResponsePayloadBuilder` calls `ConfigToApplyFetcher` that calls
`ConfigToApplyBuilder` and `OldDesiredConfigGenerator` and performs
`#diff` on both. It logs the diff in case it is not empty but returns
the older config_to_apply in both the cases.

All of this is done so that we can test out the new logic in
Create::DesiredConfig::Main::main to create
desired_config.