Fix pagination headers for "references a commit is pushed to" API
The "references a commit is pushed to" API was not returning correct
offset pagination headers in the response. `Link` did not contain a
"next" entry even when there were more pages available, and
`x-next-page` was blank.

This was because we were providing Kaminari with only enough records
to respond to the given request, which resulted in Kaminari behaving
as if there were no more pages available.

This was solved by giving Kaminari one extra record so that it knows
there are more records available and accordingly sets the headers.
For example, if 100 records were requested, we fetch 101 records, so
that there is one more record than required.

When we actually reach the end of the data and there are no more
pages, Kaminari will see that and set headers correctly.