Filter GraphQL nodes based on the `@introduced` directive
Currently, when deploying GitLab, we can't guarantee the order of the
deployment between ~backend and ~frontend. To avoid problems of the
frontend depending on new GraphQL fields, we ensure to introduce the
~backend changes on a major release before the fields are required on
the ~frontend. However, this introduces a hard lock on feature
development having to wait ~1 month between ~backend and ~frontend
development.

To avoid that, we're introducing the `@introduced(milestone: VALUE)`
directive on our GraphQL. This way, fields/objects/fragments on the
client side (~frontend) can specify when the field/object/fragment was introduced,
       and we filter out these on the ~backend if they are planned for a future
       version of the ~backend.

       This will streamline the development of features enabling the ~frontend
       and ~backend to be delivered on the same milestone, even if the
       ~frontend deploys happens before the ~backend deployment.