Enable transform-logical-assignment-operators globally
Due to an outdated version of webpack we have to enable
@babel/plugin-transform-logical-assignment-operators
locally for several imported packages.

Instead of that, we can simply enable this plugin globally,
like we have done for optional chaining and nullish coalescing
operator to simplify our build configuration.

Additionally this cleans up other plugins that are already
enabled globally.