Hide default options in filter while searching
When filtering, if the user has entered a search term, do not show the
default options in the filter token dropdown.

For example, if a user is filtering by label, when they search for a
label name, do not show the "None" and "Any" options above the search
results.

Changelog: changed