Avoid cross-project pipeline artifacts in gitlab-backup-cli
Previously `gitlab-backup-cli` required artifacts to be
present for the `db:setup pg16` job in `master`, but this often
failed on forks because those artifacts don't exist.

We observe that the `db:setup pg16` only loads the schema and does not
actually insert data. We can simply download the schema directly and
avoid this cross-project dependency.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/554732