Fix 500 from too many dots in inputs expression
A CI inputs expression with a valid access followed by an invalid access
was resulting in a 500 because we correctly retrieved the value for the
valid access (often a string) and then tried to retrieve another value
from the valid value using the invalid access, resulting in a
NoMethodError from calling key? on a String (etc).

This commit fixes the 500 by ensuring that we only try to access values
that respond to `key?`. It's difficult to communicate this error to
users, so we reuse the existing error stating that an unknown input name
was provided. This error provides the user with the name of the invalid
access which should help with debugging.

Changelog: fixed
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/434299